Refactor Decryption Mediator
break out the decryption segment into smaller files.  move and rename refactor GuardianDataStore to generic DataStore
