+++ src/electionguard/decrypt_with_secrets.py
@@ -1,6 +1,4 @@
-from typing import Dict, List, Optional, Tuple, Union
-from multiprocessing import cpu_count
-from multiprocessing.dummy import Pool
+from typing import List, Optional
 
 from .ballot import (
     CiphertextBallot,
@@ -10,110 +8,21 @@ from .ballot import (
     PlaintextBallotContest,
     PlaintextBallotSelection,
 )
-from .decryption_share import (
-    CiphertextDecryptionSelection,
-    DecryptionShare,
-    get_tally_shares_for_selection,
-)
-from .dlog import discrete_log
 from .election import (
     InternalElectionDescription,
     ContestDescriptionWithPlaceholders,
     SelectionDescription,
 )
-from .group import ElementModP, ElementModQ, mult_p, div_p
+from .group import ElementModP, ElementModQ
 from .hash import hash_elems
 from .logs import log_warning
 from .nonces import Nonces
 
-# from .process import GlobalProcessingPool, CPU_POOL
-from .tally import (
-    CiphertextTallyContest,
-    PlaintextTallyContest,
-    CiphertextTallySelection,
-    PlaintextTallySelection,
-)
-from .types import CONTEST_ID, GUARDIAN_ID, SELECTION_ID
 from .utils import get_optional
 
 ELECTION_PUBLIC_KEY = ElementModP
 
-# TODO: split this into two files, one decrypt with secrets, the other decrypt with guardians
-
-CiphertextSelection = Union[CiphertextBallotSelection, CiphertextTallySelection]
-
-
-def decrypt_selection_with_decryption_shares(
-    selection: CiphertextSelection,
-    shares: Dict[
-        GUARDIAN_ID, Tuple[ELECTION_PUBLIC_KEY, CiphertextDecryptionSelection]
-    ],
-    extended_base_hash: ElementModQ,
-    suppress_validity_check: bool = False,
-) -> Optional[PlaintextTallySelection]:
-    """
-    Decrypt the specified `CiphertextTallySelection` with the collection of `ElementModP` decryption shares
-
-    :param selection: a `CiphertextTallySelection`
-    :param shares: the collection of shares to decrypt the selection
-    :param extended_base_hash: the extended base hash code (𝑄') for the election
-    :return: a `PlaintextTallySelection` or `None` if there is an error
-    :param suppress_validity_check: do not validate the encryption prior to decrypting (useful for tests)
-    """
-
-    if not suppress_validity_check:
-        # Verify that all of the shares are computed correctly
-        for guardian_id, share in shares.items():
-            public_key, decryption = share
-            # verify we have a proof or recovered parts
-            if decryption.proof is None and decryption.recovered_parts is None:
-                log_warning(
-                    f"decrypt selection failed for guardian: {guardian_id} selection: {selection.object_id} with missing data"
-                )
-                return None
-
-            if decryption.proof is not None and decryption.recovered_parts is not None:
-                log_warning(
-                    f"decrypt selection failed for guardian: {guardian_id} selection: {selection.object_id} cannot have proof and recovery"
-                )
-                return None
-
-            if decryption.proof is not None and not decryption.proof.is_valid(
-                selection.message, public_key, decryption.share, extended_base_hash,
-            ):
-                log_warning(
-                    f"decrypt selection failed for guardian: {guardian_id} selection: {selection.object_id} with invalid proof"
-                )
-                return None
-
-            if decryption.recovered_parts is not None:
-                for (
-                    compensating_guardian_id,
-                    part,
-                ) in decryption.recovered_parts.items():
-                    if not part.proof.is_valid(
-                        selection.message,
-                        part.recovery_key,
-                        part.share,
-                        extended_base_hash,
-                    ):
-
-                        log_warning(
-                            f"decrypt selection failed for guardian: {guardian_id} selection: {selection.object_id} with invalid partial proof"
-                        )
-                        return None
-
-    # accumulate all of the shares calculated for the selection
-    all_shares_product_M = mult_p(
-        *[decryption.share for (_, decryption) in shares.values()]
-    )
-
-    # Calculate 𝑀=𝐵⁄(∏𝑀𝑖) mod 𝑝.
-    decrypted_value = div_p(selection.message.beta, all_shares_product_M)
-    d_log = discrete_log(decrypted_value)
-    return PlaintextTallySelection(
-        selection.object_id, d_log, decrypted_value, selection.message,
-    )
+# The Methods in this file can be used to decrypt values if private keys or nonces are known
 
 
 def decrypt_selection_with_secret(
@@ -312,91 +221,6 @@ def decrypt_contest_with_nonce(
             return None
 
     return PlaintextBallotContest(contest.object_id, plaintext_selections)
-
-
-def decrypt_tally_contests_with_decryption_shares_async(
-    tally: Dict[CONTEST_ID, CiphertextTallyContest],
-    shares: Dict[GUARDIAN_ID, DecryptionShare],
-    extended_base_hash: ElementModQ,
-) -> Optional[Dict[CONTEST_ID, PlaintextTallyContest]]:
-    """
-    Decrypt the specified tally within the context of the specified Decryption Shares
-
-    :param tally: the encrypted tally of contests
-    :param shares: a collection of `DecryptionShare` used to decrypt
-    :param extended_base_hash: the extended base hash code (𝑄') for the election
-    :return: a collection of `PlaintextTallyContest` or `None` if there is an error
-    """
-    contests: Dict[CONTEST_ID, PlaintextTallyContest] = {}
-
-    cpu_pool = Pool(cpu_count())
-
-    # iterate through the tally contests
-    for contest in tally.values():
-        selections: Dict[SELECTION_ID, PlaintextTallySelection] = {}
-
-        plaintext_selections = cpu_pool.starmap(
-            decrypt_selection_with_decryption_shares,
-            [
-                (
-                    selection,
-                    get_tally_shares_for_selection(selection.object_id, shares),
-                    extended_base_hash,
-                )
-                for (_, selection) in contest.tally_selections.items()
-            ],
-        )
-
-        # verify the plaintext values are received and add them to the collection
-        for plaintext in plaintext_selections:
-            if plaintext is None:
-                log_warning(f"could not decrypt tally for contest {contest.object_id}")
-                return None
-            selections[plaintext.object_id] = plaintext
-
-        contests[contest.object_id] = PlaintextTallyContest(
-            contest.object_id, selections
-        )
-
-    cpu_pool.close()
-
-    return contests
-
-
-def decrypt_tally_contests_with_decryption_shares(
-    tally: Dict[CONTEST_ID, CiphertextTallyContest],
-    shares: Dict[GUARDIAN_ID, DecryptionShare],
-    extended_base_hash: ElementModQ,
-) -> Optional[Dict[CONTEST_ID, PlaintextTallyContest]]:
-    """
-    Decrypt the specified tally within the context of the specified Decryption Shares
-
-    :param tally: the encrypted tally of contests
-    :param shares: a collection of `DecryptionShare` used to decrypt
-    :param extended_base_hash: the extended base hash code (𝑄') for the election
-    :return: a collection of `PlaintextTallyContest` or `None` if there is an error
-    """
-    contests: Dict[CONTEST_ID, PlaintextTallyContest] = {}
-
-    # iterate through the tally contests
-    for contest in tally.values():
-        selections: Dict[SELECTION_ID, PlaintextTallySelection] = {}
-
-        for (_, selection) in contest.tally_selections.items():
-            tally_shares = get_tally_shares_for_selection(selection.object_id, shares)
-            plaintext_selection = decrypt_selection_with_decryption_shares(
-                selection, tally_shares, extended_base_hash
-            )
-            if plaintext_selection is None:
-                log_warning(f"could not decrypt tally for contest {contest.object_id}")
-                return None
-            selections[plaintext_selection.object_id] = plaintext_selection
-
-        contests[contest.object_id] = PlaintextTallyContest(
-            contest.object_id, selections
-        )
-
-    return contests
 
 
 def decrypt_ballot_with_secret(
